% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cNMA_b.R
\name{cNMA_b}
\alias{cNMA_b}
\title{Perform Bayesian cNMA using rjags}
\usage{
cNMA_b(df, range, maxna, ns, model, nc)
}
\arguments{
\item{df}{Dataframe for analysis}

\item{range}{Columns with the values and components}

\item{maxna}{Maximum number of arms}

\item{ns}{Number of studies}

\item{model}{Model for cNMA}

\item{nc}{Number of components}
}
\value{
Results of Bayesian cNMA
}
\description{
Perform Bayesian cNMA using rjags
}
