% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cNMA_f.R
\name{cNMA_f}
\alias{cNMA_f}
\title{Perform Frequentist cNMA using netmeta}
\usage{
cNMA_f(df, range, maxna, ns, chkident = FALSE)
}
\arguments{
\item{df}{Dataframe for analysis}

\item{range}{Columns with the components}

\item{maxna}{Maximum number of arms}

\item{ns}{Number of studies}

\item{chkident}{Boolean for details.chkident}
}
\value{
Results of Frequentist cNMA
}
\description{
Perform Frequentist cNMA using netmeta
}
