% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMA_f.R
\name{NMA_f}
\alias{NMA_f}
\title{Perform Frequentist NMA using netmeta}
\usage{
NMA_f(df, maxna, ns)
}
\arguments{
\item{df}{Dataframe for analysis}

\item{maxna}{Maximum number of arms}

\item{ns}{Number of studies}
}
\value{
sth
}
\description{
Perform Frequentist NMA using netmeta
}
\examples{
data(example)
NMA_f(example, 3, 5)
}
