% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_f.R
\name{reformat_f}
\alias{reformat_f}
\title{Reformat the data for frequentist cNMA}
\usage{
reformat_f(df, range, maxna, ns)
}
\arguments{
\item{df}{Dataframe to reformat}

\item{range}{Columns to reformat (components only)}

\item{maxna}{Maximum number of arms}

\item{ns}{Number of studies}
}
\value{
Dataframe for Frequentist cNMA
}
\description{
Reformat the data for frequentist cNMA
}
