% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_nma.R
\name{reformat_nma}
\alias{reformat_nma}
\title{Reformat the data for Frequentist NMA}
\usage{
reformat_nma(df, maxna, ns)
}
\arguments{
\item{df}{Dataframe to reformat}

\item{maxna}{Maximum number of arms}

\item{ns}{Number of studies}
}
\value{
Dataframe for Frequentist NMA
}
\description{
Reformat the data for Frequentist NMA
}
\examples{
data(example)
reformat_nma(example, 3, 5)
}
