% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_b.R
\name{reformat_b}
\alias{reformat_b}
\title{Reformat the data for bayesian cNMA}
\usage{
reformat_b(df, maxna, ns)
}
\arguments{
\item{df}{Dataframe to reformat}

\item{maxna}{Maximum number of arms}

\item{ns}{Number of studies}
}
\value{
List of variables for Bayesian cNMA
}
\description{
Reformat the data for bayesian cNMA
}
